+++ CHANGELOG.md
@@ -5,9 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0] - 2026-02-11
+## [Unreleased]
 
 ### Added
+
+- approximate conversion bounds for mixed state with standard protocol.
+
+### Fixed
+
+- [#4](https://github.com/qat-inria/stellar-numerics/pull/4) fixed typing issues by making `StellarProfile` covariant with the type of their state.
+
+## [0.1] - 2026-02-11
+
+### Added
+
 First commit.
 
 ### Fixed